wip undo
WIP implementation was trying to have an undo stack for each active
instance of animus, and then, when getting the `onChanged` event
from chrome.storage.sync, check if it's a change or an instance
doing undo.

However, without uglier hacks, it's not possible to tell whether
a createX+deleteX is different from a createX+undo.

The way to go is to have a single undo stack which is shared on
sync storage, which solves some of the ugly case scenarios.