Fix animus/li witdh
Closes #89, "make li size have a max-width". Depending on the window's
width, the entries were running off from the left.

Closes #44, "list item width should be 100% of animus-view". Since
`li`'s weren't full width, there were 'holes' between entries that
caused the input to be focused while scrolling (since mouseout causes
the input to be focused).

Animus width is now calculated via the max width of the content of all
entries. The width of the content of an entry is now calculated via
rendering its text to a canvas.